% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_mat.R
\name{bin_mat}
\alias{bin_mat}
\title{Bin values in a matrix}
\usage{
bin_mat(X, breaks)
}
\arguments{
\item{X}{A named matrix containing values between 0 and 100 (e.g. \code{LECCDSExPkg::test_mat}).}

\item{breaks}{A vector of doubles containing the break values to use when binning X (See LECCDSExPkg::break_options).}
}
\value{
The input matrix
}
\description{
Bin values in a matrix using a set of breaks.
}
\examples{
mat <- LECCDSExPkg::bin_mat(X = LECCDSExPkg::test_mat, breaks = LECCDSExPkg::break_options[["20_inc"]])
}
