% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{break_options}
\alias{break_options}
\title{A list of breaks for binning data}
\format{
A named list of vectors containing 4 items.
}
\usage{
break_options
}
\description{
A named list of breaks for binning data between 0 and 100 (e.g. \code{LECCDSExPkg::test_mat}) into categories.
}
\details{
\code{break_options}
}
\keyword{datasets}
