% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming_distance.R
\name{hamming_distance}
\alias{hamming_distance}
\title{Calculate the hamming distance}
\usage{
hamming_distance(X)
}
\arguments{
\item{X}{A named matrix containing values between 0 and 100 (e.g. \code{LECCDSExPkg::test_mat}), with the values usually binned using \code{LECCDSExPkg::bin_mat}.}
}
\value{
A matrix containing the hamming distance values between each pairwise combination of column values in X.
}
\description{
Calculate the hamming distance between each pairwise combination of column values in X.
}
\examples{
mat <- LECCDSExPkg::bin_mat(X = LECCDSExPkg::test_mat, breaks = LECCDSExPkg::break_options[["50_inc"]])
LECCDSExPkg::hamming_distance(X = mat)
}
