% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_mat}
\alias{test_mat}
\title{A matrix of random values between 0 and 100}
\format{
A matrix containing 10 and 20
}
\usage{
test_mat
}
\description{
A matrix of random values between 0 and 100 with additional values randomly assigned to 0.
}
\details{
\code{test_mat}
}
\keyword{datasets}
